#29 - HomeController 삭제 및 CommonController 수정
HomeController와 CommonController 를 통합
- 단순 페이지 접근 url 들은 CommonController
- sign up 기능은 ProfileController 로 이동
